fix: blob driver
